+++ runtime/common/Cargo.toml
@@ -105,6 +105,11 @@ git = 'https://github.com/paritytech/polkadot'
 branch = 'apopiak-xcm-origin'
 default-features = false
 
+[dependencies.xcm-executor]
+git = 'https://github.com/paritytech/polkadot'
+branch = 'apopiak-xcm-origin'
+default-features = false
+
 # cumulus
 [dependencies.cumulus-primitives-core]
 git = 'https://github.com/paritytech/cumulus.git'
